add 3_2, red 3.1

+++ Razdel_3/Lesson_1/Lesson3_1.cpp
@@ -10,7 +10,7 @@
 
 int main () {
     bool stopcode;
-    int m,n;
+    long m,n;
     std::cout << "m=";
     std::cin >> m;
     std::cout << "n=";
